+++ changelog.md
@@ -1,3 +1,9 @@
+# Version 1.0.2
+
+### Added CLI Command Maker
+_ run `npm run createCommand` to Automatically create a command file with the provided answers
+
+
 # Version 1.0.1
 
 ### Added ErrorEmbed.js